Added `BigInteger` to `Neo.Extensions`
