+++ src/Neo.Extensions/BigIntegerExtensions.cs
@@ -1,0 +1,66 @@
+// Copyright (C) 2015-2024 The Neo Project.
+//
+// BigIntegerExtensions.cs file belongs to the neo project and is free
+// software distributed under the MIT software license, see the
+// accompanying file LICENSE in the main directory of the
+// repository or http://www.opensource.org/licenses/mit-license.php
+// for more details.
+//
+// Redistribution and use in source and binary forms with or without
+// modifications are permitted.
+
+using System;
+using System.Numerics;
+using System.Runtime.CompilerServices;
+
+namespace Neo.Extensions
+{
+    internal static class BigIntegerExtensions
+    {
+        internal static int GetLowestSetBit(this BigInteger i)
+        {
+            if (i.Sign == 0)
+                return -1;
+            var b = i.ToByteArray();
+            var w = 0;
+            while (b[w] == 0)
+                w++;
+            for (var x = 0; x < 8; x++)
+                if ((b[w] & 1 << x) > 0)
+                    return x + w * 8;
+            throw new Exception();
+        }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        internal static BigInteger Mod(this BigInteger x, BigInteger y)
+        {
+            x %= y;
+            if (x.Sign < 0)
+                x += y;
+            return x;
+        }
+
+        internal static BigInteger ModInverse(this BigInteger a, BigInteger n)
+        {
+            BigInteger i = n, v = 0, d = 1;
+            while (a > 0)
+            {
+                BigInteger t = i / a, x = a;
+                a = i % x;
+                i = x;
+                x = d;
+                d = v - t * x;
+                v = x;
+            }
+            v %= n;
+            if (v < 0) v = (v + n) % n;
+            return v;
+        }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        internal static bool TestBit(this BigInteger i, int index)
+        {
+            return (i & (BigInteger.One << index)) > BigInteger.Zero;
+        }
+    }
+}
