+++ src/Neo.Extensions/Neo.Extensions.csproj
@@ -15,7 +15,7 @@
   <ItemGroup>
     <InternalsVisibleTo Include="Neo" />
     <InternalsVisibleTo Include="Neo.IO" />
-    <InternalsVisibleTo Include="$(AssemblyName).Tests" />
+    <InternalsVisibleTo Include="Neo.UnitTests" />
   </ItemGroup>
 
 </Project>
