+++ src/Neo/Cryptography/ECC/ECPoint.cs
@@ -9,6 +9,7 @@
 // Redistribution and use in source and binary forms with or without
 // modifications are permitted.
 
+using Neo.Extensions;
 using Neo.IO;
 using Neo.IO.Caching;
 using System;
