+++ tests/Neo.UnitTests/UT_Helper.cs
@@ -11,6 +11,7 @@
 
 using FluentAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using Neo.Extensions;
 using Neo.IO.Caching;
 using Neo.Network.P2P;
 using Neo.SmartContract;
